+++ src/xdoc/node.rs
@@ -52,11 +52,7 @@ impl Node {
 
     /// Returns true if this node is either a Node::Text or a Node::CData.
     pub fn is_text(&self) -> bool {
-        match self {
-            Node::Text(_) => true,
-            Node::CData(_) => true,
-            _ => false,
-        }
+        matches!(self, Node::Text(_) | Node::CData(_))
     }
 }